+++ src/components/Layout/DashBoardLayout2/index.tsx
@@ -1,7 +1,7 @@
 import DashBoardHeader from '@/components/Header/DashBoard';
 import SideMenu from '@/components/SideMenu';
 import { atom } from 'jotai';
-import { PropsWithChildren, useState } from 'react';
+import { PropsWithChildren } from 'react';
 
 export const sideMenuAtom = atom(false);
